+++ backendGigaBank/src/main/java/br/com/guilhermeRibeiro/backendGigaBank/entity/ContaBancaria.java
@@ -4,6 +4,7 @@ import br.com.guilhermeRibeiro.backendGigaBank.dto.request.contaBancaria.ContaBa
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "t_contaBancaria")
@@ -21,7 +22,7 @@ public class ContaBancaria {
     @Column(length = 6)
     private String numero;
 
-    private Double saldo;
+    private BigDecimal saldo;
 
     @NotNull
     @ManyToOne
@@ -64,11 +65,11 @@ public class ContaBancaria {
         this.numero = numero;
     }
 
-    public Double getSaldo() {
+    public BigDecimal getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(Double saldo) {
+    public void setSaldo(BigDecimal saldo) {
         this.saldo = saldo;
     }
